+++ Bot/Libs/kumiko_ui_components/__init__.py
@@ -1,9 +1,11 @@
+from .modals import QuestsCreateModal, QuestsDeleteOneModal, QuestsUpdateTimeModal
 from .views import (
     AHPurgeAllView,
     ALPurgeDataView,
     CreateAccountView,
     MarketplacePurgeAllView,
     PurgeAccountView,
+    QuestsDeleteOneConfirmView,
     QuestsPurgeAllView,
 )
 
@@ -14,4 +16,8 @@ __all__ = [
     "QuestsPurgeAllView",
     "CreateAccountView",
     "PurgeAccountView",
+    "QuestsDeleteOneModal",
+    "QuestsDeleteOneConfirmView",
+    "QuestsCreateModal",
+    "QuestsUpdateTimeModal",
 ]